+++ nym-vpn-core/crates/nym-gateway-directory/src/lib.rs
@@ -7,6 +7,7 @@ mod entries;
 mod error;
 pub mod gateway_cache;
 mod gateway_client;
+mod gateway_store;
 mod helpers;
 
 pub use nym_sdk::mixnet::{NodeIdentity, Recipient};